Gameplay_Errors_Corrected
Update to correct an error where if the ball vector starts in the
direction of the computer's paddle, the input variable into its fuzzy
system was undefined and it threw an error. Now it starts correctly!
Also added one more re-eval of logic since it was weirdly missing the
ball later in the game? Hopefully that fixes things but if not I'll look
into ensuring the "offensive" moves using the edge of the paddle aren't
thinking the paddle edge coordinates are way off where they actually
are.