Created an Executable
Created an executable version of the singles game, available if you have
the matlab compiler runtime. There's a randomized flattening function on
the ball to make it harder for the computer to win, but now I'm scoring
all the time so I might re-upload this without it.  Also the ball
movement is super laggy but I'm not sure if it's my dinosaur of a
laptop.... will test on a friend's laptop!